chore(Gruntfile): adds grunt-contrib-watch support
this commit introduces a new grunt-contrib-watch plugin to
let a watcher run in the background when developing on
angular-leap. Currently, it just executes the jshint
task that was introduced a commit before this commit.

I'll expand the watch task by adding karma tasks as well,
to run all provided tests when a change on any source file
occurs. But thats up next :)