chore(docs): adds ngdocs support
this commit introduces the support for ngdocs to generate
some fancy documentation based on our source code. :)

ngdocs is used by the angular core itself and is also used
in other known projects such as angular-scrolly or angular-translate.

you can now run grunt ngdocs which will generate a site folder
for you. once this is generated you just have to start a simple
http server there to see the result. :)